+++ src/apis/mutations/useDeletePostMutation.ts
@@ -26,22 +26,12 @@ const deletePost = async ({ postId }: CustomRequestData) => {
 
 const useDeletePostMutation = (postId: string) => {
   const queryClient = useQueryClient();
-  const toast = useToast();
-  const navigate = useNavigate();
 
   return useMutation<PostResponse, AxiosError, CustomRequestData>({
     mutationFn: () => deletePost({ postId }),
     onSuccess: (data) => {
       queryClient.invalidateQueries(queryKey.posts.all);
       queryClient.removeQueries(queryKey.posts.detail(data._id));
-
-      toast({
-        title: '머쓱이가 결국 떠났어요.',
-        status: 'success',
-        position: 'top'
-      });
-
-      navigate('/');
     }
   });
 };
